refactor: split rev methods from issuer
